requirements.txt
google-api-python-client
vaderSentiment
emoji
python-dotenv
matplotlib
